requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
backports.tarfile==1.2.0
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
dulwich==0.22.8
factory_boy==3.3.3
Faker==36.2.2
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
kombu==5.4.2
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pbs-installer==2025.2.12
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
RapidFuzz==3.12.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.3.3.18
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.5
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
zipp==3.21.0
zstandard==0.23.0
